Fix dead link in 1.4->1.5 upgrade documentation
[ upstream commit 8a76414fb19fd8ec57a5b92df6591056d246c898 ]

The broken link was caused by using the 'latest' URL domain and path as
root. \ |SCM_WEB|\/examples/kubernetes/1.10/cilium-pre-flight-with-rm-svc-v2.yaml
Therefore the URL picked up the most recent tags, and post 1.5.12 the
path `/examples/kubernetes/1.X` were removed, so all newly built URLs were broken.

This change makes the URL static to the release path of `1.5.12`. If
there are other options to keep the URL tagged within the `1.5.x` minor
versions; I'm interested in learning about it, but this is a quick fix.

Fixes: #10391

Signed-off-by: Joshua Roppo <joshroppo@gmail.com>
Signed-off-by: vagrant <vagrant@runtime1>